test(TaxForm): Make sure to trigger if form was not submitted for this year
